Separate utils and testing tests into different files
